价格改double类型

+++ core-model/src/main/java/com/toutiao/app/domain/plot/PlotDetailsFewDo.java
@@ -103,12 +103,12 @@ public class PlotDetailsFewDo {
      * 均价
      */
     @ChangeName("averagePrice")
-    private Integer avgPrice;
+    private Double avgPrice;
     /**
      * 总价
      */
     @ChangeName("totalPrice")
-    private Integer sumPrice;
+    private Double sumPrice;
     /**
      * 建成年代
      */
